+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/event/PaymentFailedEvent.java
@@ -1,4 +1,4 @@
-package com.loopers.domain.payment;
+package com.loopers.domain.payment.event;
 
 public record PaymentFailedEvent(Long orderId, Long userId, Long couponId, String reason) {
 }
